# Fix performance regression with JSON Modules