This PR fixes the performance regression introduced in #1065 

It appears that trying to "trick" the compiler by swapping compiler options caused TypeScript to do a whole load of reallocation of internals, etc.  Just ignoring the "invalid" error appears to be a better scenario.

My local benchmarks:

|Benchmark|0fbee30 (Before #1065)|6345b60 (Current Master)|eb7104d (This PR)|
|--|--|--|--|
|Cold 002|1.337s|1.895s|1.320s|
|Cold 003|1.396s|2.089s|1.427s|